Moves JunkPickupDialog and ProvideAddress intents out to their own files.
