+++ lib/alexa-skill/index.js
@@ -2,18 +2,20 @@
  * @module      com.louisvillemetro.the-city
  * @author      Reydel Leon Machado
  * @copyright   Copyright (c) 2016 Reydel Leon Machado
- * @license     This code is license under the MIT license. See the LICENSE file in the project root for license terms.
+ * @license     This code is licensed under the MIT license. See the LICENSE file in the project root for license terms.
  */
 
 "use strict";
 
-var Config     = require('../../config'),
-    config     = new Config(),
-    util       = require('util'),
-    AlexaSkill = require('./alexa-skill'),
-    Promise    = require('bluebird'),
-    request    = require('request-promise'),
-    _          = require('lodash');
+var Config                       = require('../../config'),
+    config                       = new Config(),
+    util                         = require('util'),
+    AlexaSkill                   = require('./alexa-skill'),
+    Promise                      = require('bluebird'),
+    request                      = require('request-promise'),
+    _                            = require('lodash'),
+    handleJunkPickUpDialogIntent = require('./intent-handlers/junk-pickup-dialog'),
+    handleProvideAddressIntent   = require('./intent-handlers/provide-address');
 
 // ----------------------- Define the skill -----------------------
 
@@ -128,112 +130,6 @@ function handleSessionEndRequest(response) {
 
         resolve(response.tell(speechOutput, null));
     });
-}
-
-function handleJunkPickUpDialogIntent(event, response) {
-    return new Promise(function (resolve) {
-        let speechOutput = {
-            speech: `Sure! Tell me your address, number first, including your city or state. Ignore the unit number and zip code.`,
-            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-        };
-
-        let repromptSpeech = {
-            speech: `Please, tell me your address, number first?`,
-            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-        };
-
-        resolve(response.ask(speechOutput, repromptSpeech));
-    });
-}
-
-function handleProvideAddressIntent(event, response) {
-    let streetAddress = event.request.intent.slots.StreetAddress;
-
-    if (streetAddress && streetAddress.value) {
-        let googleApiUrl = config.googleGeoApi.url + config.googleGeoApi.format + '?address=' + encodeURI(streetAddress.value) + '&key=' + config.googleGeoApi.key;
-
-        return request.get(googleApiUrl)
-                .then(function (verifiedAddress) {
-                    verifiedAddress = JSON.parse(verifiedAddress);
-
-                    var speechOutput,
-                        repromptSpeech;
-
-                    if (verifiedAddress.status && verifiedAddress.status == 'OK') {
-                        speechOutput = {
-                            speech: `Thanks! Just to make sure: Is your address ${verifiedAddress.results[0].formatted_address}?`,
-                            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                        };
-
-                        repromptSpeech = {
-                            speech: `Please, confirm if your address is ${verifiedAddress.results[0].formatted_address}`,
-                            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                        };
-
-                        // Segment address for Junk Pickup API consumption
-                        var verifiedStreetNumber  = '',
-                            verifiedStreetAddress = '',
-                            verifiedCity          = '';
-
-                        _.forEach(verifiedAddress.results[0].address_components, function (component) {
-                            if (component.types.includes('street_number')) {
-                                verifiedStreetNumber = component.long_name;
-                            }
-
-                            if (component.types.includes('route')) {
-                                verifiedStreetAddress = component.short_name;
-                            }
-
-                            if (component.types.includes('locality')) {
-                                verifiedCity = component.long_name;
-                            }
-                        });
-
-                        response._session.attributes.VerifiedStreetAddress = verifiedStreetNumber + ' ' + verifiedStreetAddress;
-                        response._session.attributes.VerifiedCity = verifiedCity;
-                    } else {
-                        speechOutput = {
-                            speech: `Uhmmm! The address you provided seems to be invalid. Please, repeat your address, number first. Remember to include your city or state.`,
-                            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                        };
-
-                        repromptSpeech = {
-                            speech: `Please, repeat your address, number first.`,
-                            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                        };
-                    }
-
-                    return response.ask(speechOutput, repromptSpeech);
-                }).catch(function (err) {
-                    console.dir(err);
-
-                    let speechOutput = {
-                        speech: `I'm having some trouble verifying that address. Please, repeat your address, number first.`,
-                        type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                    };
-
-                    let repromptSpeech = {
-                        speech: `Please, repeat your address, number first`,
-                        type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                    };
-
-                    return response.ask(speechOutput, repromptSpeech);
-                });
-    } else {
-        return new Promise(function (resolve) {
-            let speechOutput = {
-                speech: `I didn't quite get that. Please, repeat your address, number first.`,
-                type: AlexaSkill.speechOutputType.PLAIN_TEXT
-            };
-
-            let repromptSpeech = {
-                speech: `Please, repeat your address, number first.`,
-                type: AlexaSkill.speechOutputType.PLAIN_TEXT
-            };
-
-            resolve(response.ask(speechOutput, repromptSpeech));
-        });
-    }
 }
 
 function handleAmazonYesIntent(event, response) {
